@@ -7,7 +7,7 @@ maintainer_email = stefan.appelhoff@mailbox.org
 description = A Python implementation of the preprocessing pipeline (PREP) for EEG data.
 long-description = file: README.rst
 long-description-content-type = text/x-rst; charset=UTF-8
-keywords = EEG artifact preprocessing data
+keywords = EEG artifact preprocessing data neuroscience neuroimaging mne python electroencephalography eeg
 license = MIT
 license_files = LICENSE
 platforms = any
@@ -27,11 +27,10 @@ project_urls =
   Source = https://github.com/sappelhoff/pyprep
 
 [options]
-python_requires = >= 3.6
+python_requires = >= 3.7
 install_requires =
   numpy >= 1.14.1
   scipy >= 1.0.0
-  statsmodels >= 0.8.0
   mne >= 0.20.0
   psutil >= 5.4.3
 packages = find:
@@ -67,9 +66,13 @@ exclude =
   *docs/auto_examples/*
   *build/lib*
   examples/
+  *matprep_artifacts/*
 
 [isort]
-skip_glob=examples/*
+profile=black
+skip_glob =
+  examples/*
+  matprep_artifacts/*
 
 [coverage:run]
 omit =
