@@ -5,7 +5,7 @@ Run the full PREP
 
 
 In this example we show how to run PREP with ``pyprep``. We also compare
-:class:`prep_pipeline.PrepPipeline` with PREP's results in Matlab.
+:class:`~pyprep.PrepPipeline` with PREP's results in Matlab.
 
 We use sample EEG data from Physionet EEG Motor Movement/Imagery Dataset:
 `https://physionet.org/content/eegmmidb/1.0.0/ <https://physionet.org/content/eegmmidb/1.0.0/>`_
